fix: 修复代理设置审查问题
- proxyPort UserDefaults 存取统一用 Int 类型，与项目模式一致
- 添加端口范围警告提示（低于 1024 自动调整）
- 删除空 MARK extension 冗余代码
- proxy caption 明确标注绑定地址 127.0.0.1
- 中英文本地化同步更新

+++ Sources/CodexBar/SettingsStore+Defaults.swift
@@ -676,12 +676,13 @@ extension SettingsStore {
         }
     }
 
+    /// 代理端口（UInt16），UserDefaults 以 Int 存取，与项目其他数值字段一致。
     var proxyPort: UInt16 {
         get { self.defaultsState.proxyPort }
         set {
-            let clamped = max(1024, min(65535, newValue))
+            let clamped = UInt16(max(1024, min(65535, Int(newValue))))
             self.defaultsState.proxyPort = clamped
-            self.userDefaults.set(clamped, forKey: "proxyPort")
+            self.userDefaults.set(Int(clamped), forKey: "proxyPort")
             CodexBarLog.logger(LogCategories.settings).info(
                 "Proxy port updated",
                 metadata: ["port": "\(clamped)"])
@@ -710,6 +711,4 @@ extension SettingsStore {
         }
         return self.normalizeProviders(providers, maxCount: maxCount)
     }
-
-    // MARK: - Proxy
 }
